+++ EubosChess/src/main/java/eubos/position/IEvaluate.java
@@ -1,5 +1,5 @@
 package eubos.position;
 
 public interface IEvaluate {
-	int evaluatePosition(IPositionAccessors pos);
+	short evaluatePosition(IPositionAccessors pos);
 }
