+++ EubosChess/src/main/java/eubos/position/IScoreMate.java
@@ -3,5 +3,5 @@ package eubos.position;
 import eubos.board.pieces.Piece.Colour;
 
 public interface IScoreMate {
-	int scoreMate(int currPly, boolean isWhite, Colour initialOnMove);
+	short scoreMate(byte currPly, boolean isWhite, Colour initialOnMove);
 }
