+++ EubosChess/src/main/java/eubos/position/MateScoreGenerator.java
@@ -15,34 +15,34 @@ public class MateScoreGenerator implements IScoreMate {
 		this.searchDepth = searchDepth;	
 	}
 	
-	public int scoreMate(int currPly, boolean isWhite, Colour initialOnMove) {
+	public short scoreMate(byte currPly, boolean isWhite, Colour initialOnMove) {
 		// Handle mates (indicated by no legal moves)
-		int mateScore = 0;
+		short mateScore = 0;
 		if (pos.isKingInCheck()) {
 			mateScore = generateScoreForCheckmate(currPly);
 			// If white got mated, need to back up a large negative score (good for black)
 			if (isWhite)
-				mateScore=-mateScore;
+				mateScore=(short) -mateScore;
 			//debug.printMateFound(currPly);
 		} else {
 			mateScore = getScoreForStalemate();
 			// TODO: introduce a more sophisticated system for handling stalemate scoring.
 			if (initialOnMove==Colour.black)
-				mateScore=-mateScore;
+				mateScore=(short) -mateScore;
 		}
 		return mateScore;
 	}	
 	
-	private int getScoreForStalemate() {
+	private short getScoreForStalemate() {
 		// Avoid stalemates by giving them a large penalty score.
 		return -King.MATERIAL_VALUE;
 	}	
 	
-	int generateScoreForCheckmate(int currPly) {
+	short generateScoreForCheckmate(byte currPly) {
 		// Favour earlier mates (i.e. Mate-in-one over mate-in-three) by giving them a larger score.
 		int totalMovesSearched = searchDepth/PLIES_PER_MOVE;
 		int mateMoveNum = (currPly-1)/PLIES_PER_MOVE; // currPly-1 because mate was caused by the move from the previousPly
 		int multiplier = totalMovesSearched-mateMoveNum;
-		return multiplier*King.MATERIAL_VALUE;
+		return (short)(multiplier*King.MATERIAL_VALUE);
 	}	
 }
