+++ EubosChess/src/main/java/eubos/search/IMoveGenerator.java
@@ -8,6 +8,6 @@ import eubos.board.InvalidPieceException;
 
 interface IMoveGenerator {
 	public SearchResult findMove() throws NoLegalMoveException, InvalidPieceException;
-	public SearchResult findMove(int searchDepth) throws NoLegalMoveException, InvalidPieceException;
-	public SearchResult findMove(int searchDepth, LinkedList<GenericMove> lastPc) throws NoLegalMoveException, InvalidPieceException;
+	public SearchResult findMove(byte searchDepth) throws NoLegalMoveException, InvalidPieceException;
+	public SearchResult findMove(byte searchDepth, LinkedList<GenericMove> lastPc) throws NoLegalMoveException, InvalidPieceException;
 }
